kustomize: Add isolated overlay
The default kustomize overlay at "config/overlay/default" needs to populate a
RoleBinding under
kube-system and a ClusterRole and ClusterRoleBinding this mean that some
resources have to be populated out of the tenant namespace on infra
cluster. This change add another kustomize overlay "config/overlay/isolated"
that prevent that by disabling external apiserver authentication
(/metrics and /configz endpoint will not be valid) and disabling zone
and region and tenant nodes. Also the "config/overlay/kubevirtci" points
now to isolated.

Signed-off-by: Enrique Llorente <ellorent@redhat.com>