Allow podman to push images to the local insecure registry
THe kubevirt CI infrastructure is moving to using a podman based
bootstrap image rather than a docker in docker setup.

When running the pull-kubevirt-cloud-provider-kubevirt-e2e presubmit
with podman, the job fails[1] as podman doe not allow pusgin to insecure
registries by default[2].

This update will add the local insecure registry to registries.conf when
the podman socket is available which will allow podman to push images to
the registry.

[1] https://prow.ci.kubevirt.io/view/gs/kubevirt-prow/pr-logs/pull/kubevirt_project-infra/2359/rehearsal-pull-kubevirt-cloud-provider-kubevirt-e2e/1575140184981245952#1:build-log.txt%3A766
[2] https://github.com/containers/podman/blob/main/troubleshooting.md#4-http-server-gave-http-response-to-https-client

Signed-off-by: Brian Carey <bcarey@redhat.com>